JobAssitanceAI/requirements.txt
anthropic>=0.25.0
python-docx>=1.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0
playwright>=1.40.0
python-dotenv>=1.0.0